adds Pedro Rodrigues @pedro2555 signature
Most likelly for nothing. This was called out 4 years ago and yielded to
tangible results, some time was spent in public relations non sense but
the end result is the same. That time I moved to GitLab, and its a way
better experience.

Everyone that signed should do the same, and STOP using GitHub. That is
the only way to really show your lack of support.

And this is not a thing with GitHub, it is a Microsoft thing. If you
didn't feel it comming at the time, I sure did. And the first ICE
contract was my bail out moment.

What are you waiting for?